dom/media/test/test_decode_error_crossorigin.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Invalid formats raise decode errors with default messages for CORS cross-origin media</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">
var manager = new MediaTestManager;
var gotErrSrcNotSupported = false;
var gotErrDecode = false;

function startTest(test, token) {
  const ok = function(condition, name) {
    SimpleTest.ok(condition, `${test.name}: ${name}`);
  }
  const is = function(a, b, name) {
    SimpleTest.is(a, b, `${test.name}: ${name}`);
  }

  const v = document.createElement("video");
  manager.started(token);
  v.addEventListener("error", event => {
    if (v.readyState == v.HAVE_NOTHING) {
      is(v.error.code, MediaError.MEDIA_ERR_SRC_NOT_SUPPORTED);
      is(v.error.message, "Failed to open media");
      gotErrSrcNotSupported = true;
    } else {
      is(v.error.code, MediaError.MEDIA_ERR_DECODE);
      is(v.error.message, "Failed to decode media");
      gotErrDecode = true;
    }
    manager.finished(token);
  });

  v.addEventListener("loadeddata", () => {
    ok(false, "Unexpected loadeddata event");
    manager.finished(token);
  });

  v.addEventListener("ended", () => {
    ok(false, "Unexpected ended event");
    manager.finished(token);
  });

  v.autoplay = true;

  // CORS-cross-origin URL.
  v.src = `http://example.com/tests/dom/media/test/${test.name}`;
}

SpecialPowers.pushPrefEnv({
  "set": [
    ["media.cache_size", 40000],
  ]
}, () => {
  manager.onFinished = () => {
    ok(gotErrSrcNotSupported, "At least one test led to src-not-supported");
    ok(gotErrDecode, "At least one test led to a decode error");
  };
  manager.runTests(gDecodeErrorTests, startTest);
});
</script>
</pre>
</body>
</html>
